requirements.txt
Django==5.0.2
djangorestframework==3.14.0
python-dotenv==1.0.1
django-cors-headers==4.3.1
django-redis==5.4.0
mysqlclient==2.2.4
django-filter==23.5
djangorestframework-simplejwt==5.3.1 